# docs/api/api-specification.yaml
openapi: 3.0.0
info:
  title: Genesis Music API
  description: AI-powered guitar learning platform API
  version: 1.0.0
  contact:
    name: Genesis Music Team
    email: api@genesismusic.com
servers:
  - url: https://api.genesismusic.com/v1
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /transcribe:
    post:
      tags: [Transcription]
      summary: Start transcription job
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audioFile:
                  type: string
                  format: binary
                  description: Audio file (MP3, WAV, M4A)
                youtubeUrl:
                  type: string
                  format: url
                  description: YouTube video URL
                options:
                  type: object
                  properties:
                    tuning:
                      type: string
                      enum: [standard, drop_d, half_step_down, open_g]
                      default: standard
                    difficulty:
                      type: string
                      enum: [beginner, intermediate, advanced]
                      default: intermediate
                    includeTheory:
                      type: boolean
                      default: true
      responses:
        202:
          description: Transcription job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, processing]
                  estimatedTime:
                    type: integer
                    description: Estimated processing time in seconds
        400:
          $ref: '#/components/responses/BadRequest'
        429:
          description: Rate limit exceeded
  /transcribe/{jobId}:
    get:
      tags: [Transcription]
      summary: Get transcription status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionJob'
        404:
          description: Job not found

    delete:
      tags: [Transcription]
      summary: Cancel transcription job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Job cancelled
        404:
          description: Job not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TranscriptionJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        sourceType:
          type: string
          enum: [youtube, upload]
        sourceUrl:
          type: string
        sourceFilename:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          properties:
            duration:
              type: integer
              description: Duration in seconds
            tempo:
              type: integer
              description: BPM
            keySignature:
              type: string
            timeSignature:
              type: string
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...